requirements.txt
blinker
certifi
charset-normalizer
click
dnspython
flask
Flask-Cors
google-api-core
google-api-python-client
google-apps-meet
google-auth
google-auth-httplib2
google-auth-oauthlib
googleapis-common-protos
grpcio
grpcio-status
httplib2
idna
importlib-metadata
itsdangerous
jinja2
MarkupSafe
numpy
oauthlib
pandas
proto-plus
protobuf
pyasn1
pyasn1-modules
pymongo
pyparsing
python-dateutil
python-dotenv
pytz
requests
requests-oauthlib
rsa
six
tzdata
uritemplate
urllib3
werkzeug
zipp

# Web Frameworks
flask-socketio
gunicorn
eventlet

# Database & Caching
redis

# Social Media APIs
tweepy
python-instagram
linkedin-api

# NLP & AI Libraries 
textblob 
transformers 
torch 
nltk 
emoji

# Message Queue 
celery

# Notifications 
twilio 
sendgrid

# Visualization 
plotly 
dash

# Utilities 
pydantic 

# Testing 
pytest

# Security 
PyJWT 
cryptography 
